#Android App Template 1) Restructuring
