+++ app/src/main/java/com/example/androidapplicationtemplate/data/repositoryImpl/SomeRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.example.androidapplicationtemplate.domain.repositoryImpl
+package com.example.androidapplicationtemplate.data.repositoryImpl
 
 import com.example.androidapplicationtemplate.data.local.localDataSource.SomeLocalDataSource
 import com.example.androidapplicationtemplate.data.remote.remoteDataSource.SomeRemoteDataSource
-import com.example.androidapplicationtemplate.data.repository.SomeRepository
+import com.example.androidapplicationtemplate.domain.repository.SomeRepository
 import javax.inject.Inject
 
 class SomeRepositoryImpl @Inject constructor(
