+++ app/src/main/java/com/example/androidapplicationtemplate/domain/repository/SomeRepository.kt
@@ -1,4 +1,4 @@
-package com.example.androidapplicationtemplate.data.repository
+package com.example.androidapplicationtemplate.domain.repository
 
 interface SomeRepository{
 	suspend fun someCrudOperation() : Result<String>
