+++ app/src/main/java/com/example/androidapplicationtemplate/domain/usecase/SomeUseCase.kt
@@ -1,6 +1,6 @@
 package com.example.androidapplicationtemplate.domain.usecase
 
-import com.example.androidapplicationtemplate.data.repository.SomeRepository
+import com.example.androidapplicationtemplate.domain.repository.SomeRepository
 import javax.inject.Inject
 
 class SomeUseCase @Inject constructor(
